job-search/serivces/Job-recommender/scraping_module/requirements.txt
selenium
request
json
pymongo